pyproject.toml
[project]
name = "reviewer-agent"
version = "1.0.0"
description = "A specialized AI agent for reviewing academic papers using the A2A protocol"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "a2a-sdk>=0.3.0",
    "uvicorn>=0.17.6",
    "starlette>=0.19.1",
    "PyPDF2>=2.10.0",
    "pdfplumber>=0.7.0",
    "arxiv>=1.4.2",
    "semanticscholar>=0.4.0",
    "httpx>=0.23.0",
    "openai>=1.0.0",
    "python-dotenv>=0.20.0",
    "pydantic>=1.9.1",
    "tenacity>=8.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true